# Fixed index counting bug in attribution