--- a/setup.py
+++ b/setup.py
@@ -7,7 +7,7 @@ with open("README.md", "r") as fh:
 
 setuptools.setup(
     name="selfies",
-    version="2.1.0",
+    version="2.1.1",
     author="Mario Krenn, Alston Lo, and many other contributors",
     author_email="mario.krenn@utoronto.ca, alan@aspuru.com",
     description="SELFIES (SELF-referencIng Embedded Strings) is a "