I've revised the indexing added in 2.1.0 to follow SMILES string index exactly and SELFIES token index exactly. Previously it used the current length of the SMILES string list which could contain empty elements. In SELFIES, it incorrectly treated order of index. This PR also increments version/updates changelog. 

New behavior:

```py
SELFIES [C][N][C][Branch1][C][P][C][C][Ring1][=Branch1]
SMILES C1NC(P)CC1
Attribution:
AttributionMap(index=0, token='C', attribution=[Attribution(index=0, token='[C]')])
AttributionMap(index=2, token='N', attribution=[Attribution(index=1, token='[N]')])
AttributionMap(index=3, token='C', attribution=[Attribution(index=2, token='[C]')])
AttributionMap(index=5, token='P', attribution=[Attribution(index=3, token='[Branch1]'), Attribution(index=5, token='[P]')])
AttributionMap(index=7, token='C', attribution=[Attribution(index=6, token='[C]')])
AttributionMap(index=8, token='C', attribution=[Attribution(index=7, token='[C]')])
SELFIES [C][N][C][C][Branch1][P][C][C][Ring1][=Branch1]
SMILES C1NC(P)CC1
Attribution:
AttributionMap(index=0, token='[C]', attribution=[Attribution(index=0, token='C')])
AttributionMap(index=1, token='[N]', attribution=[Attribution(index=2, token='N')])
AttributionMap(index=2, token='[C]', attribution=[Attribution(index=3, token='C')])
AttributionMap(index=5, token='[P]', attribution=[Attribution(index=5, token='P')])
AttributionMap(index=3, token='[C]', attribution=[Attribution(index=5, token='P')])
AttributionMap(index=4, token='[Branch1]', attribution=[Attribution(index=5, token='P')])
AttributionMap(index=6, token='[C]', attribution=[Attribution(index=7, token='C')])
AttributionMap(index=7, token='[C]', attribution=[Attribution(index=8, token='C')])
AttributionMap(index=8, token='[Ring1]', attribution=None)
AttributionMap(index=9, token='[=Branch1]', attribution=None)
```