@MarioKrenn6240 2022-06-10
> @MarioKrenn6240 @akshat998 - It looks like all the dataset tests failed because the SMILES couldn't parse correctly. Maybe we need to pin a dependency here?

we are on it, thanks!