@whitead 2022-07-11
Hi @alstonlo - you have a sharp eye! I did not spot that line switch - thanks. 

For the second comment about timing, I did try what you said and found it to add quite a bit of complexity to the code because there are possible empty tokens added that require if/else statements for a counter approach. I decided to just join them (usually quite fast) to count instead. 

I've put timing information below. Some paths are slower/faster on PR so I don't think it's a meaningful impact on performance. 

master 0f0061a
```sh
16.10s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path1]
10.49s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path9]
9.98s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path11]
9.93s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path3]
6.23s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path14]
6.11s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path13]
4.94s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path10]
3.60s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path2]
2.23s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path4]
1.98s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path5]
1.86s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path12]
1.58s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path6]
0.62s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path7]
0.24s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path8]
0.04s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path0]
```

PR [e51e3dc](https://github.com/aspuru-guzik-group/selfies/pull/89/commits/e51e3dc433aa1842669a2564281982f54d08c269)
```sh
15.74s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path1]
10.62s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path9]
10.04s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path11]
9.71s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path3]
6.45s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path14]
5.91s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path13]
4.80s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path10]
3.54s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path2]
2.27s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path4]
2.03s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path5]
1.86s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path12]
1.60s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path6]
0.62s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path7]
0.24s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path8]
0.04s call     tests/test_on_datasets.py::test_roundtrip_translation[test_path0]
```